fix that grep was broken (!)
Yes, there's a bug in grep in FreeBSD!  You might be amused that
there is a bug in the grep implementation in the FreeBSD source code.
It doesn't really matter for them, since they build with things
[here](https://github.com/freebsd/freebsd-src/blob/d88828148d893e3d7c747f9331187ec6e44f26bd/usr.bin/grep/util.c#L535)
when `WITH_INTERNAL_NOSPEC` is defined they only check for a match
in the one case `grepbehave == GREP_FIXED,` and silently just
let everything match when grepbehave is anything else.  That
grepbehave is only GREP\_FIXED when you run grep via `fgrep.`  In BSD
they have a non\-POSIX regexec with extra features so they don't
need to use WITH\_INTERNAL\_NOSPEC, so I guess they just don't care
or ever test that the code is wrong in that case.   Anyway, I was
wondering why grep on [cowasm.sh](http://cowasm.sh) was totally
broken, and didn't expect it to be because of such a significant bug.
I am not going to ever report this bug upstream because the process
for reporting or fixing a bug in FreeBSD is way too complicated.